# api/openapi.yaml
openapi: 3.0.4
info:
  title: Favorite Game Gauntlet
  version: 0.9.0
paths:
...
  /users/self/display-name:
    post:
      operationId: changeOwnDisplayName
      tags: [ Users ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Name'
      responses:
        '204':
          description: Success.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getOwnDisplayName
      tags: [ Users ]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Name'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: DisplayName not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/infos:
    get:
      operationId: getUserInfos
      tags: [ Users ]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfos'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/self/stats:
    get:
      operationId: getOwnStats
      tags: [ Users ]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/stats:
    get:
      operationId: getUserStats
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/territory-point-history:
    get:
      operationId: getUserTerritoryPointHistory
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerritoryPointChanges'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/free-point-history:
    get:
      operationId: getUserFreePointHistory
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreePointChanges'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/wishlist-games:
    get:
      operationId: getUserWishlistGames
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistGames'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/game-history:
    get:
      operationId: getUserGameHistory
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Games'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{login}/wheel-effect-history:
    get:
      operationId: getUserWheelEffectHistory
      tags: [ Users ]
      parameters:
        - $ref: "#/components/parameters/UserLogin"
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolledWheelEffects'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    UserLogin:
      name: login
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UserLogin'
  schemas:
...
    UserLogin:
      type: string
      pattern: '^\w+$'
      minLength: 3
      maxLength: 35
    Name:
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 70
...
    Duration:
      type: string
      format: duration
      example: PT2h10m5s
      description: The duration notation as defined by ISO 8601
    Points:
      type: integer
      format: int
      minimum: 0
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [ code, message ]
...
    UserStats:
      type: object
      properties:
        territoryPoints:
          $ref: '#/components/schemas/Points'
        freePoints:
          $ref: '#/components/schemas/Points'
        availableRolls:
          $ref: '#/components/schemas/Points'
        territoryHours:
          $ref: '#/components/schemas/Points'
        experiencePoints:
          $ref: '#/components/schemas/Points'
      required: [ territoryPoints, freePoints, availableRolls ]
    UserInfo:
      allOf:
        - type: object
          properties:
            login:
              $ref: '#/components/schemas/UserLogin'
            displayName:
              $ref: '#/components/schemas/Name'
          required: [ displayName ]
        - $ref: '#/components/schemas/UserStats'
    UserInfos:
      type: array
      items:
        $ref: '#/components/schemas/UserInfo'
    WishlistGame:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
      required: [ name ]
    WishlistGames:
      type: array
      items:
        $ref: '#/components/schemas/WishlistGame'
    Game:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        state:
          type: string
          enum: [ 'started', 'finished', 'cancelled' ]
        timeSpent:
          $ref: '#/components/schemas/Duration'
        finishDate:
          type: string
          format: date-time
      required: [ name, state, timeSpent ]
    Games:
      type: array
      items:
        $ref: '#/components/schemas/Game'
...
    WheelEffect:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        description:
          type: string
      required: [ name ]
...
    RolledWheelEffect:
      allOf:
        - $ref: '#/components/schemas/WheelEffect'
        - type: object
          properties:
            position:
              type: integer
              format: int
            rollDate:
              type: string
              format: date-time
          required: [ rollDate ]
    RolledWheelEffects:
      type: array
      items:
        $ref: '#/components/schemas/RolledWheelEffect'
    VictoryPointChange:
      type: object
      properties:
        changeSource:
          type: string
        changeValue:
          type: integer
          format: int
        actualChangeValue:
          type: integer
          format: int
        finalValue:
          $ref: '#/components/schemas/Points'
        changeDate:
          type: string
          format: date-time
      required: [ changeSource, changeValue, finalValue, changeDate ]
    TerritoryPointChange:
      allOf:
        - $ref: '#/components/schemas/VictoryPointChange'
        - type: object
          properties:
            changeSource:
              enum: [ 'territory-obtaining', 'territory-loss', 'other' ]
    TerritoryPointChanges:
      type: array
      items:
        $ref: '#/components/schemas/TerritoryPointChange'
    FreePointChange:
      allOf:
        - $ref: '#/components/schemas/VictoryPointChange'
        - type: object
          properties:
            changeSource:
              enum: [ 'own-wheel-effect', 'other-wheel-effect', 'quest', 'base-teleport', 'sandstorm', 'other' ]
            wheelEffectName:
              $ref: '#/components/schemas/Name'
    FreePointChanges:
      type: array
      items:
        $ref: '#/components/schemas/FreePointChange'